packages/vector/moss/pyproject.toml
[project]
name = "cognee-community-vector-adapter-moss"
version = "0.1.0"
description = "Moss vector database adapter for cognee"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "moss>=1.0.0",
    "cognee>=1.0.0",
    "python-dotenv>=0.19.0"
]
